src/Aurochses.IdentityServer.Website/Controllers/HomeController.cs
using Aurochses.IdentityServer.Website.Filters;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Hosting;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Logging;

namespace Aurochses.IdentityServer.Website.Controllers
{
    [SecurityHeaders]
    [AllowAnonymous]
    public class HomeController : Controller
    {
        private readonly ILogger _logger;
        private readonly IHostingEnvironment _hostingEnvironment;

        public HomeController(ILogger<HomeController> logger, IHostingEnvironment hostingEnvironment)
        {
            _logger = logger;
            _hostingEnvironment = hostingEnvironment;
        }

        public IActionResult Index(string returnUrl)
        {
            if (_hostingEnvironment.IsDevelopment())
            {
                // only show in development
                return View();
            }

            _logger.LogInformation("Homepage is disabled in production. Redirect to SignIn page.");
            return RedirectToAction("Index", "SignIn", new { ReturnUrl = returnUrl });
        }
    }
}
